chore: move useCookieUserConfig to new composable subfolder
